Refactor tests even more
